[AB-035] Rename a package
Package documentation was also added.
