+++ pom.xml
@@ -110,9 +110,9 @@
 
         <dependency>
             <groupId>org.codehaus.groovy</groupId>
-            <artifactId>groovy-all</artifactId>
-            <version>2.4.6</version>
-            <!--<scope>provided</scope>-->
+            <artifactId>groovy</artifactId>
+            <version>2.4.7</version>
+            <scope>provided</scope>
         </dependency>
 
         <!-- This works with dagger 2.1 (now only snapshot) -->
