+++ src/main/resources/META-INF/services/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,3 +1,3 @@
 moduleName=autobuilder-module
 moduleVersion=${version}
-extensionClasses=com.github.jakubkolar.autobuilder.extensions.BuilderDSLGroovyMethods
+extensionClasses=com.github.jakubkolar.autobuilder.groovy.BuilderDSLGroovyMethods
