+++ src/test/groovy/com/github/jakubkolar/autobuilder/groovy/VariableTest.groovy
@@ -1,0 +1,49 @@
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016 Jakub Kolar
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package com.github.jakubkolar.autobuilder.groovy
+
+import nl.jqno.equalsverifier.EqualsVerifier
+import spock.lang.Specification
+
+class VariableTest extends Specification {
+
+    def "It is a value object"() {
+        expect:
+        EqualsVerifier.forClass(Variable).verify()
+    }
+
+    def "It allows to resolve nested properties as new variables"() {
+        given:
+        def x = new Variable('x')
+
+        when:
+        def nested = x.y.z
+
+        then:
+        assert nested == new Variable('x.y.z')
+        assert nested.getName() == 'x.y.z'
+    }
+
+}
